При удалении трека, загружается диалог При удалении трека происходит удаление неиспользуемых треков из базы данных

+++ app/src/main/java/com/playlistmaker/data/db/playlist/musicTrackList/TrackListDao.kt
@@ -1,13 +1,11 @@
 package com.playlistmaker.data.db.playlist.musicTrackList
 
 import androidx.room.Dao
-import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import com.playlistmaker.data.db.favourite.MusicTrackEntity
 import kotlinx.coroutines.flow.Flow
-import retrofit2.http.DELETE
 
 @Dao
 interface TrackListDao {
@@ -23,5 +21,5 @@ interface TrackListDao {
     suspend fun deleteAll()
 
     @Query("DELETE FROM all_tracks_table WHERE id = :idToDelete")
-    fun deleteTrackById(idToDelete:Int)
+    fun deleteTrackById(idToDelete:Long)
 }
